Add TF-IDF baseline, fix fp16/fp32 crash, measure math macro-F1
Measured on the math test set (473 responses, 57 held-out dialogues):

  (a)  majority class      mean macro-F1 0.253
  (a2) TF-IDF + logreg     mean macro-F1 0.517   ECE 0.076

Per-dimension for (a2): MI 0.544 [0.495, 0.604], ML 0.509, PG 0.502,
AC 0.515. Shortcut gaps are small in both directions, worst -0.12 on
providing-guidance length, i.e. the model tracks response length slightly
less than the human labels do rather than more.

New: experiments/linear_baseline.py, wired into run_experiments.py as a2.
It runs in 8 seconds and answers a question neither the majority floor nor
the transformer does -- how much of this task falls out of surface lexical
features. It also gives a real trained reference point on hardware that
cannot host a DeBERTa run.

Bug fix, models/modeling.py: the deberta-v3-base checkpoint is stored in
fp16 and transformers 5.x honors the checkpoint dtype instead of upcasting,
so the encoder came back fp16 while the heads stayed fp32. That surfaced
twice -- "mat1 and mat2 must have the same dtype" on CPU, and a Metal
assertion on MPS that aborted the process rather than raising. The encoder
is now loaded with dtype=torch.float32 explicitly. resolve_device() in
models/train.py centralizes device choice and honors TGT_DEVICE.

run_c now treats a missing programming dataset as a skip, so the in-domain
math number can be produced before any programming labeling exists.

configs/base.yaml drops to batch_size 8 with gradient_accumulation_steps 2
(effective batch unchanged at 16). Batch 16 at 512 tokens peaks high enough
to swap on a 16GB machine, which slowed training roughly 30x without ever
erroring.

Not done: the DeBERTa math run itself. It is blocked on hardware, not code
-- an isolated forward/backward runs at 1.7 s/step on MPS, but this machine
had 7.2GB of 8GB swap already in use and the full run projected to 29 hours.
The model path stays covered by the 5 slow tests.

+++ models/modeling.py
@@ -40,7 +40,12 @@ class MultiHeadGrader(nn.Module):
         self.n_classes = n_classes
 
         self.config = AutoConfig.from_pretrained(model_name)
-        self.encoder = AutoModel.from_pretrained(model_name)
+        # Force fp32. The deberta-v3-base checkpoint is stored in fp16, and
+        # transformers 5.x honors the checkpoint dtype instead of upcasting, so
+        # the default gives an fp16 encoder feeding fp32 heads: a "mat1 and mat2
+        # must have the same dtype" crash on CPU, and a Metal assertion on MPS.
+        # Fine-tuning wants fp32 master weights anyway.
+        self.encoder = AutoModel.from_pretrained(model_name, dtype=torch.float32)
         hidden = self.config.hidden_size
         self.dropout = nn.Dropout(dropout)
         self.heads = nn.ModuleList([nn.Linear(hidden, n_classes) for _ in range(n_dimensions)])
